Extracted Axes class from widget-base and renamed it PrototypeAxes
This also moved the remaining compiler warnings from widget-base to the new widget-prototype-axes file.
